chore: move login and register dto classes to separate packages
